Change playRound return to object initializer
The only thing that was returned was a binary value indicating whether
the player had won the round or not.

The object initializer allows storing data from each single round to be
reused elsewere (eg. for diplaying end of round/game stats).

The initialized properties are:
result: [`Win` | `Lose`] A string representing the outcome of the round
winner: [computerSelection | playerSelection] The winning shapes value
loser:  [computerSelection | playerSelection] The losing shapes value