Refactor playRound and playGame to integrate UI
Most of the game variables are now global.

The loop in the playGame function is no longer necessary.

The playRound function is now global and gets called by each button
with the displayed shape as the playerSelection argument.

The remaining code of the playGame function are the checks for ending
the game, so they are now the checkWinner function that gets called
near the end of each playRound call.